+++ library/objective-c/EnvoyNetworkMonitor.m
@@ -4,16 +4,8 @@
 
 #import <Foundation/Foundation.h>
 #import <Network/Network.h>
-#import <SystemConfiguration/SystemConfiguration.h>
 
 @implementation EnvoyNetworkMonitor
-
-+ (void)startReachabilityIfNeeded {
-  static dispatch_once_t reachabilityStarted;
-  dispatch_once(&reachabilityStarted, ^{
-    _start_reachability();
-  });
-}
 
 + (void)startPathMonitorIfNeeded {
   static dispatch_once_t monitorStarted;
@@ -22,7 +14,7 @@
   });
 }
 
-#pragma mark - Private (Network Path Monitor)
+#pragma mark - Private
 
 static nw_path_monitor_t _path_monitor;
 
@@ -52,9 +44,6 @@ static void _start_path_monitor() {
     NSLog(@"[Envoy] setting preferred network to %@", isCellular ? @"WWAN" : @"WLAN");
     set_preferred_network(isCellular ? ENVOY_NET_WWAN : ENVOY_NET_WLAN);
 
-    // TODO(jpsim): Should we shadow or otherwise compare these results with the reachability
-    // flags?
-
     // TODO(jpsim): Should we report back other properties of the reachable path?
     //
     // - nw_path_get_status:
@@ -78,42 +67,6 @@ static void _start_path_monitor() {
   });
 
   nw_path_monitor_start(_path_monitor);
-}
-
-#pragma mark - Private (Reachability)
-
-static SCNetworkReachabilityRef _reachability_ref;
-
-static void _reachability_callback(SCNetworkReachabilityRef target,
-                                   SCNetworkReachabilityFlags flags, void *info) {
-  if (flags == 0) {
-    return;
-  }
-
-  BOOL isUsingWWAN = flags & kSCNetworkReachabilityFlagsIsWWAN;
-  NSLog(@"[Envoy] setting preferred network to %@", isUsingWWAN ? @"WWAN" : @"WLAN");
-  set_preferred_network(isUsingWWAN ? ENVOY_NET_WWAN : ENVOY_NET_WLAN);
-}
-
-static void _start_reachability() {
-  NSString *name = @"io.envoyproxy.envoymobile.EnvoyNetworkMonitor";
-  SCNetworkReachabilityRef reachability =
-      SCNetworkReachabilityCreateWithName(nil, [name UTF8String]);
-  if (!reachability) {
-    return;
-  }
-
-  _reachability_ref = reachability;
-
-  SCNetworkReachabilityContext context = {0, NULL, NULL, NULL, NULL};
-  if (!SCNetworkReachabilitySetCallback(_reachability_ref, _reachability_callback, &context)) {
-    return;
-  }
-
-  dispatch_queue_t queue = dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0);
-  if (!SCNetworkReachabilitySetDispatchQueue(_reachability_ref, queue)) {
-    SCNetworkReachabilitySetCallback(_reachability_ref, NULL, NULL);
-  }
 }
 
 @end