+++ docs/root/intro/version_history.rst
@@ -13,6 +13,8 @@ Features:
 
 - API: added Envoy's response flags to final stream intel (:issue:`#2009 <2009>`)
 - size: The size of the dynamic library was reduced by ~46% (:issue:`#2053 <2053>`)
+- iOS: Removed ``EngineBuilder.enableNetworkPathMonitor(_:)``, now ``NWPathMonitor`` is always
+  used to observe network reachability state.
 
 0.4.5 (January 13, 2022)
 ========================